fix: secp256k1 ecdsa mul - fix handling of point at infinity (#16679)
Fixes the issue found by @federicobarbacovi in `secp256k1_ecdsa_mul`.
The issue was that the point at infinity flag `_is_infinity` was not
being set explicitly when adding base points due to the skew flags. This
issue arises when in the following edge case:

### Description of the issue

Choose secp256k1 scalars $s_1, u_1, u_2 \in \mathbb{F}_r$ such that 
1. the public key is computed as $P = s_1 \cdot G$ for generator $G \in
\mathbb{G}$
2. $u_1 \cdot G + u_2 \cdot P = \mathcal{O}$
where $\mathcal{O}$ is the point at infinity.

In the `secp256k1_ecdsa_mul` function, the scalars $u_1$ and $u_2$ are
split in endomorphism scalars:

$$u_1 = u_{1, \textsf{low}} + \lambda \cdot u_{1, \textsf{high}}, \quad
u_2 = u_{2, \textsf{low}} + \lambda \cdot u_{2, \textsf{high}}.$$

Each of the scalars $u_{1, \textsf{low}}, u_{1, \textsf{high}}, u_{2,
\textsf{low}}, u_{2, \textsf{high}}$ has a skew bit: let them be $b_1,
b_2, b_3, b_4$. We handle the skew bit in the accumulator $A$ as:

1. First compute addition $B = A + G \quad \text{or} \quad B = A - G$
based on the skew being positive or negative
2. Set the accumulator to $B$ if the skew is non-zero, else keep the
accumulator unchanged: $A := (1 - b) \cdot A + b \cdot B$

While performing the second step, we conditionally selected only the `x`
and `y` coordinates of $A$ or $B$ but didn't fetch the `_is_infinity`
flag. This leads to the flag `_is_infinity` set to `false` even when the
result of the scalar multiplication is point at infinity.

### Fix

The fix is to correctly set the `_is_infinity` flag when performing
conditional selection operation. So we just define a new
`conditional_select` function on the `biggroup` (and `biggroup_goblin`)
class.

---------

Co-authored-by: federicobarbacovi <federicobarbacovi@users.noreply.github.com>