fix: rewire some internal oracles
This is fairly simple: private exec had an override for `fetchTaggedLogs`, despite it doing the exact same thing as the base utility exec (since there is no difference in message processing between utility and private). Additionally, utility implemented emitOffchainMessage despite it not actually being available (and throwing). I moved it to the other interface, but did not change the iface itself as I want to do that in a single cohesive pass to avoid compatibility problems.