fix: outdated cli reference
The cli reference is extremely outdated. This does a first pass at at least bringing it up to nightly 9/26. A thing we could look into is getting the output of `aztec start --help`, then processing it and diffing it to next. Each time we see a change, we can manually edit what needs to be changed. This is a step below full automation that was discussed earlier but would be much easier and still provide utility to help us catch when we need to update relevant docs so this drift can be avoided.