refactor!: cleaning up tagging oracles
In this PR I merge `getIndexedTaggingSecretAsSender` and `incrementAppTaggingSecretIndexAsSender` and `get_app_tag_as_sender` into single `get_next_app_tag(sender, recipient)`. This results in the app tag being computed in TS and not in Noir.

The motivation for this is that to improve robustness of tagging as described in [this post](https://forum.aztec.network/t/on-note-discovery-and-index-coordination/7165) we need to associate a given index with tx hash and this is the first step in that direction.

### Note for reviewer

This is a breaking change because it changes oracles but I don't think a dev would ever call these oracles directly so I didn't bother with writing migration notes. If the reviewer thinks it should be done then LMK and I can add it.

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>