fix!: tx trace should handle reverts early in phase followed by skipped items (#17411)
## The bug
TX trace was failing proving when a transaction reverted early in a
phase and skipped later entries in that same phase. For example:
1. App logic has 2 enqueued calls
2. First one reverts
3. Second is skipped

The `TX_READ_PHASE_LENGTH` lookup to public inputs was failing because
tracegen was using the `phase.events` length to populate the
`remaining_phase_counter` column, but that wasn't right when there was
an early revert in the phase! In that case, we would write the "number
of events emitted for the phase" into this column, and then the lookup
to public inputs would fail.

Furthermore, we had a constraint that `remaining_phase_counter` must be
1 at the end of the phase. That constraint needed to caveat it with
"except if reverted" (`* (1 - reverted)`).

## What i did
1. Included the `phase_lengths` (total number of items that COULD be
processed in a phase) in `TxStartupEvent`s to properly tracegen this
column for all phases
2. Constrained that `end_phase` gets 1 when `reverted == 1` (and misc
related cleanup including constraining that `reverted` is bool).
3. Added bb-prover tests that replicate a ccouple variations of this