feat: `ExecutionTaggingIndexCache`
Fixes https://github.com/AztecProtocol/aztec-packages/issues/15753

In this PR I implement an ExecutionTaggingIndexCache which stores the used indexes during an execution and at the end we store them in the database. By storing that info once execution is done we'll be able to have indexes associated with a transaction hash. This is a necessary step towards improving of unconstrained tagging as described in [this post](https://forum.aztec.network/t/on-note-discovery-and-index-coordination/7165).

Given that once this PR is merged we will store the indexes in the database only for executions and not for simulations this PR should fix https://github.com/AztecProtocol/aztec-packages/issues/15753.