fix!: Include proposer signature in attestations (#17721)
Fixes A-126

- The `BlockAttestation` struct in the p2p network now requires an
additional field.

No, this was not written by Claude!

Adds the signature of the block proposer over the attestation, which
gets validated before accepting and rebroadcasting every attestation.
This prevents a malicious attestor from flooding the network with signed
attestations for non-existing proposals.

Fixes Joe's Attack. We should no longer run into this since the
attestation must include the proposer's signature, but if somehow the
attestation still makes it past to the validator, we add an extra check
there and we drop the attestation (and remove it from our pool).

We were using the in-memory attestation pool by default. This changes it
to using the lmdb one.

Every entry in the attestations retrieved from the block in the
`CommitteeAttestation.fromPacked` call in `data_retrieval.ts` must have
either a signature or an address, so the committee commitment can be
reconstructed. A completely empty entry here is an error.

Since we now require the proposer signature to construct a
`BlockAttestation` object, and the proposer signature is mixed with all
the other attestations, we change the return type from
`BlockAttestation` to a simple struct with the recovered or submitted
address, along with the status.