refactor: cleaning up `CallInterface` related utils
When I was looking at how to implement `self` approach in contracts I was investigating current implementation of how `ContractInterface` stubs are generated. This made me realize that a lot of utils there can be removed and it can all be made much cleaner by using new Noir features like `<type as Serialize>::N`.

When looking at the `create_stub_base` I figured that we can reuse the serialization functionality of `derive_serialize` so I refactored the relevant code out of `derive_serialize` into a separate function that can be reused between the 1 functions.

I removed a lot of code so overall I think the outcome is nice.