fix: Remove block number from p2p proposals and attestations
Block number was NOT being signed over, this meant any node could tweak
it and cause receiving nodes to consider the proposal incorrect. We fix
this by fetching the parent node using the last archive root (which
required adding new indices to the archiver) and computing the block
number by adding one to it.

As for attestations, the block number was not used at all, so we could
remove it.