refactor: cleaning up macros::functions::utils
This PR was quite a journey:

## Step 1

When trying to figure out how to rename the incorrectly named `get_fn_visibility` in a PR down the stack I realized that it's better to just drop it because:
- it was only used by `create_assert_correct_initializer_args`, `create_mark_as_initialized` and `create_init_check` functions which in turn where only called in `transfrom_private` and `transform_public`,
- these functions contained only 2 lines of code.

Because of this ^ it's better to just directly inline contents of these funcs into the `transform_*` functions.

When I finished this I checked where are the is `is_fn_private` and `is_fn_public` utils used which led me to step 2:

## Step 2
When checking when are the above mentioned functions used I realized that we use them in `create_authorize_once_check` and in the `stub_fn`. In both of these cases the code was just terrible because these functions are called from `transform_private` and `transform_public` functions hence the flow was:

`transform_private` --> `stub_fn` --> are we in private? --> if yes call `create_private_stub`

which I could replace with:

`transform_private` --> `create_private_stub`

So I dropped the `stub_fn` and just called the `create_private_stub` directly.

The similar was the case for `create_authorize_once_check` where I could simply pass in an `is_private` arg to the function and get rid of the `is_fn_private` dependency.

## Step 3
Once I did this I realized that I don't need the `EXTERNAL_REGISTRY` I introduced in the PR down the stack at all, removing a lot of the sad ugly complexity I needed to introduce there 🎉

## Future work
- The whole `noir-projects/aztec-nr/aztec/src/macros/functions` directory is very messy and I plan on move the code around into more logical places in the dir once this stack of PRs is merged.
- I think the `stub` name is non-descriptive enough so will try to come up with something better.

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>