fix: Do not verify contracts if not deployed in that run
Attempt at fixing the error from [this run](https://github.com/AztecProtocol/aztec-packages/actions/runs/18787851531/job/53610628682) where the verification failed due to a non-existing contract. In this run, CREATE_ROLLUP_CONTRACTS was set to false, so the `DEPLOY_ROLLUP_CONTRACTS_DIR` template did not run. My guess is that the `tf output` loaded the output from a different run on a different version which had this contract that no longer exists. In particular, `RewardDeploymentExtLib` exists in 2.0 but not in 2.1, the deployment that failed.