fix: fix sysbox resources
Seems like we regressed ourselves into a state where we couldn't run all the tests on the mainframe without hitting resource limits. e.g. running `yarn test` in yarn-project hits ~22k threads peak. This in itself needs investigating.
But, the container limit is 65k, so what was the issue?
Systemd has it's own limits - by default 15% the total, which is why bumping container pids helped, but not as much as i expected. This now ensures everything is maxed up the sysbox container limit.
As I've rebuilt the amis im also removing the ensure_zig stuff as its in the container now (as well as ldid).

I've had to fix a load of issues introduced by the foundry version bump. I'm unsure why the version was bumped when the existing version seemed to be working...