feat: Store txHash and blockHash in private event storage
First part of F-96 (about returning event metadata)

We want the PXE to expose some event metadata (mainly tx hash, block number, and block hash) to wallets.

I plan to tackle this in little self contained PR's. This first PR only affects internals, but not APIs: it modifies the PXE's private event store to store tx hash and block hash for later consumption.

I'm introducing a `PrivateEvent` type which for now is an implementation detail as it is swallowed by PXE. But eventually it could be lifted to the PXE API.

I'm not yet familliarized enough with criteria for types to be defined and made available by stdlib or Aztec.js, it might make sense to eventually have a re-usable `PrivateEvent` type _somewhere over there_, but one thing at a time.

There's also a very mildly annoying amount of duplication in `private_event_data_provider.test.ts` which I'm contributing to with this PR, but I'm opting to tackle that in subsequent PRs as I work on a more mature API.